Organized by the type of resource
